feat: 부모 연결 끊기
