+++ src/assets/data/profileImages.jsx
@@ -1,0 +1,21 @@
+import Profile1 from "~/assets/img/common/character/character_sol.svg";
+import Profile2 from "~/assets/img/common/character/character_moli.svg";
+import Profile3 from "~/assets/img/common/character/character_rino.svg";
+import Profile4 from "~/assets/img/common/character/character_shoo.svg";
+import Profile5 from "~/assets/img/common/character/character_doremi.svg";
+import Profile6 from "~/assets/img/common/character/character_lulu.svg";
+import Profile7 from "~/assets/img/common/character/character_pli.svg";
+import Profile8 from "~/assets/img/common/character/character_lay.svg";
+
+const availableProfiles = [
+  { id: 1, src: Profile1 },
+  { id: 2, src: Profile2 },
+  { id: 3, src: Profile3 },
+  { id: 4, src: Profile4 },
+  { id: 5, src: Profile5 },
+  { id: 6, src: Profile6 },
+  { id: 7, src: Profile7 },
+  { id: 8, src: Profile8 },
+];
+
+export default availableProfiles;
