+++ src/components/MyPage/Profile.jsx
@@ -7,25 +7,7 @@ import { getCredibility } from "../../utils/getCredibility";
 import { updateUserInfo } from "../../services/user";
 import { useSelector } from "react-redux";
 
-import Profile1 from "~/assets/img/common/character/character_sol.svg";
-import Profile2 from "~/assets/img/common/character/character_moli.svg";
-import Profile3 from "~/assets/img/common/character/character_rino.svg";
-import Profile4 from "~/assets/img/common/character/character_shoo.svg";
-import Profile5 from "~/assets/img/common/character/character_doremi.svg";
-import Profile6 from "~/assets/img/common/character/character_lulu.svg";
-import Profile7 from "~/assets/img/common/character/character_pli.svg";
-import Profile8 from "~/assets/img/common/character/character_lay.svg";
-
-const profiles = [
-  { id: 1, src: Profile1 },
-  { id: 2, src: Profile2 },
-  { id: 3, src: Profile3 },
-  { id: 4, src: Profile4 },
-  { id: 5, src: Profile5 },
-  { id: 6, src: Profile6 },
-  { id: 7, src: Profile7 },
-  { id: 8, src: Profile8 },
-];
+import profiles from "../../assets/data/profileImages";
 
 const Profile = ({ userInfo }) => {
   const [isEditing, setIsEditing] = useState(false);
@@ -61,12 +43,10 @@ const Profile = ({ userInfo }) => {
     const isValidBirth = profileData.birthDate.length === 10;
 
     const { credibility, ...dataWithoutCredibility } = profileData;
-    console.log("여기", dataWithoutCredibility);
 
     if (isValidName && isValidPhone && isValidBirth) {
       try {
         const data = await updateUserInfo(accessToken, dataWithoutCredibility);
-        console.log("didijidj", data);
         alert("정보가 업데이트 되었습니다.");
       } catch (error) {
         console.error(error);
